Refactoring dependencies initiation
