+++ src/model/user_domain.go
@@ -2,24 +2,6 @@ package model
 
 import "encoding/json"
 
-type UserDomainInterface interface {
-	GetEmail() string
-	GetFirstName() string
-	GetLastName() string
-	GetAge() int8
-	SetId(string)
-	GetId() string
-}
-
-func NewUserDomain(firstName, lastName, email string, age int8) UserDomainInterface {
-	return &userDomain{"",
-		firstName,
-		lastName,
-		email,
-		age,
-	}
-}
-
 type userDomain struct {
 	id        string
 	firstName string
